Fix ignore of small avatars
Previously, there was a logic error in the condition to ignore small
avatars. Although the pattern was properly excluded, the
`download_child_p` hook would always accept anything that wasn't
expected to be HTML: `... allowed(url, parent["url"]) or html == 0')`.

To fix this, I add an additional check in `download_child_p` before
the call to `allowed` to ensure that avatars are excluded.  The
`get_urls` hook didn't have this error, so no updates were necessary
there and the existing `allowed` call works as expected.

I also exclude 64 pixel GIF avatars to be thorough. This was already
done for the 16 pixel ones, but not the 64 pixel ones.